[new downloader] Fix storage issues

+++ platform/http_request.cpp
@@ -1,6 +1,7 @@
-#include "platform/http_request.hpp"
 #include "platform/chunks_download_strategy.hpp"
+#include "platform/http_request.hpp"
 #include "platform/http_thread_callback.hpp"
+#include "platform/platform.hpp"
 
 #include "defines.hpp"
 
@@ -277,7 +278,8 @@ class FileHttpRequest : public HttpRequest, public IHttpThreadCallback
         (void)my::DeleteFileX(m_filePath + RESUME_FILE_EXTENSION);
 
         // Rename finished file to it's original name.
-        (void)my::DeleteFileX(m_filePath);
+        if (Platform::IsFileExistsByFullPath(m_filePath))
+          (void)my::DeleteFileX(m_filePath);
         CHECK(my::RenameFileX(m_filePath + DOWNLOADING_FILE_EXTENSION, m_filePath), ());
 
         DisableBackupForFile(m_filePath);
